+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY3S9PGDYK9WT8192PCH8CPK"
+        "01KY4GW4DPREBQCWJFJ1ZTESNG"
     )
     r"""The unique identifier of the data source"""