+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-07-22T02:07:48.193Z")
+    updated: Optional[datetime] = parse_datetime("2026-07-22T08:59:49.949Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")