+++ packages/orq-rc/RELEASES.md
@@ -9619,3 +9619,13 @@ Based on:
 - [python v4.12.0-rc.50] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.50 - packages/orq-rc
+
+## 2026-07-22 09:03:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.51] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.51] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.51 - packages/orq-rc